debugged the lassonet part -- cause was AdamW

+++ src/sparsemodesnet/linalg/pod.py
@@ -9,7 +9,7 @@ def compute_pod_basis(X_np: np.ndarray, s: int = None):
     U, Sigma, Vt = np.linalg.svd(X_np, full_matrices=False)
     d, n = X_np.shape
     r = min(d, n) if s is None else min(s, min(d, n))
-    U_s     = U[:, :r].astype(np.float32)
-    Sigma_s = Sigma[:r].astype(np.float32)
-    Vt_s    = Vt[:r, :].astype(np.float32)
+    U_s     = U[:, :r]
+    Sigma_s = Sigma[:r]
+    Vt_s    = Vt[:r, :]
     return U_s, Sigma_s, Vt_s
